fix lint

+++ src/Graphics/OpenGL/Buffers/GeometryBuffer.cpp
@@ -17,7 +17,7 @@ namespace hws {
     // Normal Color Buffer (View-space normals)
     glGenTextures(1, &g_normal_);
     glBindTexture(GL_TEXTURE_2D, g_normal_);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width_, height_, 0, GL_RGBA, GL_FLOAT, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, (int)width_, (int)height_, 0, GL_RGBA, GL_FLOAT, nullptr);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
     glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, g_normal_, 0);
@@ -25,7 +25,7 @@ namespace hws {
     // Depth Buffer (Must be a texture so SSAO can sample it)
     glGenTextures(1, &g_depth_);
     glBindTexture(GL_TEXTURE_2D, g_depth_);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_DEPTH_COMPONENT, width_, height_, 0, GL_DEPTH_COMPONENT, GL_FLOAT, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_DEPTH_COMPONENT, (int)width_, (int)height_, 0, GL_DEPTH_COMPONENT, GL_FLOAT, nullptr);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
@@ -33,7 +33,7 @@ namespace hws {
     glFramebufferTexture2D(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_TEXTURE_2D, g_depth_, 0);
 
     // Tell OpenGL which color attachments we'll use of this framebuffer
-    unsigned int attachments[1] = {GL_COLOR_ATTACHMENT0};
+    unsigned int attachments[1] = {GL_COLOR_ATTACHMENT0}; // NOLINT
     glDrawBuffers(1, attachments);
 
     if(glCheckFramebufferStatus(GL_FRAMEBUFFER) != GL_FRAMEBUFFER_COMPLETE)
@@ -49,11 +49,11 @@ namespace hws {
 
     // Resize Normal Texture
     glBindTexture(GL_TEXTURE_2D, g_normal_);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width_, height_, 0, GL_RGBA, GL_FLOAT, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, (int)width_, (int)height_, 0, GL_RGBA, GL_FLOAT, nullptr);
 
     // Resize Depth Texture
     glBindTexture(GL_TEXTURE_2D, g_depth_);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_DEPTH_COMPONENT, width_, height_, 0, GL_DEPTH_COMPONENT, GL_FLOAT, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_DEPTH_COMPONENT, (int)width_, (int)height_, 0, GL_DEPTH_COMPONENT, GL_FLOAT, nullptr);
 
     glBindTexture(GL_TEXTURE_2D, 0);
   }
@@ -61,7 +61,7 @@ namespace hws {
   void GeometryBuffer::bind() const
   {
     glBindFramebuffer(GL_FRAMEBUFFER, fbo_);
-    glViewport(0, 0, width_, height_);
+    glViewport(0, 0, (int)width_, (int)height_);
   }
 
   void GeometryBuffer::unbind() const
